Insinuate a schema between and matching the CLI and factor body.
The CLI should work identically as before and is tested by making
before/after `json_dir` output and doing a `diff`.

This work is to demonstrate how to refactor such a CLI to gain a the
schema validation and provide a schema-protected general Pyhton API to
yet higher order generators.

The refactoring goes like:

- Copy paste current `click.option()` and `click.argument()` decorator
  stack into a Jsonnet file and modify syntax to match moo schema data
  structure.

- Create a `gen()` method which takes all args/opts as a dict.

- `gen()` calls `validate()` on that dict which loads the schema into
  `moo.otypes` and filters the dict through to its return value which
  would trigger any schema errors.

- The previous body of the `cli()` method is now in the `body()`
  method which is otheriwse unchanged.  It is then called from
  `gen()`.

So we have `CLI->gen()->validate()->body()`.

With this refactoring we gain a schema contract for this generator and
an API (the `.gen()` method) which may be called by other code w/out
detailed understanding of the arguments, as long as a schema-valid
object/dict is provided.